Moved IdeasSummaryDisplay to its own folder
