+++ src/assets/wise5/classroomMonitor/classroomMonitorComponents/shared/node-info/node-info.component.ts
@@ -4,7 +4,7 @@ import { Component, Input } from '@angular/core';
 import { ComponentFactory } from '../../../../common/ComponentFactory';
 import { ComponentServiceLookupService } from '../../../../services/componentServiceLookupService';
 import { ComponentTypeService } from '../../../../services/componentTypeService';
-import { IdeasSummaryComponent } from '../../../../directives/teacher-summary-display/ideas-summary.component';
+import { IdeasSummaryComponent } from '../../../../directives/teacher-summary-display/ideas-summary-display/ideas-summary.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { isMatchingPeriods } from '../../../../common/period/period';
 import { MatCardModule } from '@angular/material/card';
