+++ src/assets/wise5/directives/teacher-summary-display/ideas-summary-display/ideas-summary.component.html
