+++ src/assets/wise5/directives/teacher-summary-display/ideas-summary-display/ideas-summary.component.ts
@@ -1,21 +1,21 @@
-import { AnnotationService } from '../../services/annotationService';
+import { AnnotationService } from '../../../services/annotationService';
 import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
-import { ConfigService } from '../../services/configService';
-import { CRaterIdea } from '../../components/common/cRater/CRaterIdea';
-import { CRaterRubric } from '../../components/common/cRater/CRaterRubric';
-import { CRaterService } from '../../services/cRaterService';
-import { DialogGuidanceSummaryData } from './summary-data/DialogGuidanceSummaryData';
-import { IdeaData } from '../../components/common/cRater/IdeaData';
-import { IdeasSortingService } from '../../services/ideasSortingService';
-import { IdeasSummaryData } from './summary-data/IdeasSummaryData';
+import { ConfigService } from '../../../services/configService';
+import { CRaterIdea } from '../../../components/common/cRater/CRaterIdea';
+import { CRaterRubric } from '../../../components/common/cRater/CRaterRubric';
+import { CRaterService } from '../../../services/cRaterService';
+import { DialogGuidanceSummaryData } from '../summary-data/DialogGuidanceSummaryData';
+import { IdeaData } from '../../../components/common/cRater/IdeaData';
+import { IdeasSortingService } from '../../../services/ideasSortingService';
+import { IdeasSummaryData } from '../summary-data/IdeasSummaryData';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
-import { OpenResponseSummaryData } from './summary-data/OpenResponseSummaryData';
-import { SummaryService } from '../../components/summary/summaryService';
-import { TeacherDataService } from '../../services/teacherDataService';
-import { TeacherProjectService } from '../../services/teacherProjectService';
-import { TeacherSummaryDisplayComponent } from './teacher-summary-display.component';
+import { OpenResponseSummaryData } from '../summary-data/OpenResponseSummaryData';
+import { SummaryService } from '../../../components/summary/summaryService';
+import { TeacherDataService } from '../../../services/teacherDataService';
+import { TeacherProjectService } from '../../../services/teacherProjectService';
+import { TeacherSummaryDisplayComponent } from '../teacher-summary-display.component';
 
 @Component({
   imports: [CommonModule, MatCardModule, MatIconModule],
